HACS requirements
