--- const.py
@@ -1,2 +1,0 @@
-DOMAIN = "everlights3"
-HUBIP  = "hub_ip"
