+++ custom_components/client/api/default/delete_v_1_events_event_id.py
