+++ custom_components/client/api/default/get_v1_available.py
