+++ custom_components/client/errors.py
