+++ custom_components/client/models/zone_effect.py
