+++ custom_components/const.py
@@ -1,0 +1,2 @@
+DOMAIN = "everlights3"
+HUBIP  = "hub_ip"
